修改逻辑，让显示和逻辑分离。 修改数学库

+++ Server/GServer/MapServer/GameViews/BattlePerceptionView.cs
@@ -56,7 +56,7 @@ namespace MapServer.GameViews
         }
 
         #region views
-        private Dictionary<long, BattleElement> _AttachElements = new Dictionary<long, BattleElement>();
+        private Dictionary<int, BattleElement> _AttachElements = new Dictionary<int, BattleElement>();
 
         internal void DeAttachView(BattleElement battleElement)
         {
